Separated the constants and types
