+++ src/app/extractionCard/_components/drawerContent.tsx
@@ -1,22 +1,14 @@
 import { artifacts } from "../_constants/dummy";
 import { ArtifactCard } from "@/components/common/artifactCard";
+import { projectNameProp } from "../_constants/type";
 
-interface Props {
-  projectName: string;
-}
-
-export function ViewArtifactDrawer({ projectName }: Props) {
+export function ViewArtifactDrawer({ projectName }: projectNameProp) {
   return (
     <div className="p-6 space-y-4">
-      {/* Project Name */}
       <div className="text-[#000000] font-bold text-lg shadow-sm">
         <h1>{projectName}</h1>
       </div>
-
-      {/* <div className="border-b border-gray-300 my-4"></div> */}
-
       <p className="text-md font-bold text-[#363636] p-4">Artefacts</p>
-
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
         {artifacts.map((artifact, index) => (
           <ArtifactCard
@@ -27,8 +19,7 @@ export function ViewArtifactDrawer({ projectName }: Props) {
             author={artifact.author}
           />
         ))}
-
-        {/* More Button */}
+        
         <div className="flex justify-center items-center">
           <button
             className="px-6 py-2 text-sm font-semibold rounded-md"
