+++ src/app/extractionCard/_components/tabNavigation.tsx
@@ -1,20 +1,5 @@
 import React from 'react';
-
-export interface KeyValuePair {
-  key: string;
-  value: string | string[];
-}
-
-export interface ProjectCard {
-  title: string;
-  imageUrl: string;
-  keyValuePairs: KeyValuePair[];
-}
-
-export interface ProjectNavigationProps {
-  setSelectedTab: (tab: string) => void;
-  currentStatus?: string;
-}
+import { ProjectNavigationProps } from '../_constants/type';
 
 const ProjectNavigation: React.FC<ProjectNavigationProps> = ({ setSelectedTab, currentStatus}) => {
   // Function to determine active link
