+++ src/app/extractionCard/_constants/dummy.ts
